+++ packages/termansi/example/termansi_example.dart
@@ -50,6 +50,21 @@ void main() {
 
   stdout
     ..writeln()
-    ..writeln(Term.hyperLink('https://github.com/kascote/termkit', 'TermKit'))
+    ..writeln('Text Attributes:')
+    ..writeln(' ${Text.bold}Bold${Text.resetBold}')
+    ..writeln(' ${Text.dim}Dim${Text.resetDim}')
+    ..writeln(' ${Text.italic}Italic${Text.resetItalic}')
+    ..writeln(' ${Text.underline}Underline${Text.resetUnderline}')
+    ..writeln(' ${Text.doubleUnderline}Double Underline${Text.resetDoubleUnderline}')
+    ..writeln(' ${Text.curlyUnderline}Curly Underline${Text.resetCurlyUnderline}')
+    ..writeln(' ${Text.dottedUnderline}Dotted Underline${Text.resetDottedUnderline}')
+    ..writeln(' ${Text.dashedUnderline}Dashed Underline${Text.resetDashedUnderline}')
+    ..writeln()
+    ..writeln(
+        ' ${Text.doubleUnderline}${Color.underlineColor256(120)}Double Underline Color${Text.resetDashedUnderline}${Color.resetUnderlineColor}')
+    ..writeln(
+        ' ${Text.curlyUnderline}${Color.underlineTrueColor(255, 0, 0)}Curly Underline Color${Text.resetCurlyUnderline}${Color.resetUnderlineColor}')
+    ..writeln()
+    ..writeln(Term.hyperLink('https://github.com/kascote/termkit', 'Link to project'))
     ..writeln(Term.notify('TermKit', 'Hello from TermAnsi'));
 }